pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "samantha-cli"
version = "0.1.0"
description = "Give Claude a voice. A voice assistant powered by Claude, inspired by Her."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Ethan Rogers"}]
keywords = ["ai", "voice", "assistant", "claude", "tts", "speech"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "SpeechRecognition>=3.10",
    "PyAudio>=0.2.14",
    "fish-audio-sdk>=0.1",
    "pygame>=2.5",
    "rich>=13.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
samantha = "samantha.cli:main"

[project.urls]
Homepage = "https://github.com/ethanplusai/samantha-cli"
Repository = "https://github.com/ethanplusai/samantha-cli"
